loudness: merge groveloudness and hyld9767 into a single driver
