+++ examples/c++/CMakeLists.txt
@@ -68,7 +68,6 @@ add_executable (grovelinefinder-example grovelinefinder.cxx)
 add_executable (grovevdiv-example grovevdiv.cxx)
 add_executable (grovewater-example grovewater.cxx)
 add_executable (guvas12d-example guvas12d.cxx)
-add_executable (groveloudness-example groveloudness.cxx)
 add_executable (mpr121-example mpr121.cxx)
 add_executable (ublox6-example ublox6.cxx)
 add_executable (yg1006-example yg1006.cxx)
@@ -135,7 +134,7 @@ add_executable (ak8975-example ak8975.cxx)
 add_executable (lsm9ds0-example lsm9ds0.cxx)
 add_executable (eboled-example eboled.cxx)
 add_executable (mpu9250-example mpu9250.cxx)
-add_executable (hyld9767-example hyld9767.cxx)
+add_executable (loudness-example loudness.cxx)
 add_executable (mg811-example mg811.cxx)
 add_executable (wheelencoder-example wheelencoder.cxx)
 add_executable (sm130-example sm130.cxx)
@@ -190,7 +189,6 @@ include_directories (${PROJECT_SOURCE_DIR}/src/grovelinefinder)
 include_directories (${PROJECT_SOURCE_DIR}/src/grovevdiv)
 include_directories (${PROJECT_SOURCE_DIR}/src/grovewater)
 include_directories (${PROJECT_SOURCE_DIR}/src/guvas12d)
-include_directories (${PROJECT_SOURCE_DIR}/src/groveloudness)
 include_directories (${PROJECT_SOURCE_DIR}/src/mpr121)
 include_directories (${PROJECT_SOURCE_DIR}/src/ublox6)
 include_directories (${PROJECT_SOURCE_DIR}/src/yg1006)
@@ -246,7 +244,7 @@ include_directories (${PROJECT_SOURCE_DIR}/src/rgbringcoder)
 include_directories (${PROJECT_SOURCE_DIR}/src/hp20x)
 include_directories (${PROJECT_SOURCE_DIR}/src/pn532)
 include_directories (${PROJECT_SOURCE_DIR}/src/lsm9ds0)
-include_directories (${PROJECT_SOURCE_DIR}/src/hyld9767)
+include_directories (${PROJECT_SOURCE_DIR}/src/loudness)
 include_directories (${PROJECT_SOURCE_DIR}/src/mg811)
 include_directories (${PROJECT_SOURCE_DIR}/src/wheelencoder)
 include_directories (${PROJECT_SOURCE_DIR}/src/sm130)
@@ -320,7 +318,6 @@ target_link_libraries (grovelinefinder-example grovelinefinder ${CMAKE_THREAD_LI
 target_link_libraries (grovevdiv-example grovevdiv ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (grovewater-example grovewater ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (guvas12d-example guvas12d ${CMAKE_THREAD_LIBS_INIT})
-target_link_libraries (groveloudness-example groveloudness ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (mpr121-example mpr121 ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (ublox6-example ublox6 ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (yg1006-example yg1006 ${CMAKE_THREAD_LIBS_INIT})
@@ -387,7 +384,7 @@ target_link_libraries (ak8975-example mpu9150 ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (lsm9ds0-example lsm9ds0 ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (eboled-example i2clcd ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (mpu9250-example mpu9150 ${CMAKE_THREAD_LIBS_INIT})
-target_link_libraries (hyld9767-example hyld9767 ${CMAKE_THREAD_LIBS_INIT})
+target_link_libraries (loudness-example loudness ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (mg811-example mg811 ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (wheelencoder-example wheelencoder ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries (sm130-example sm130 ${CMAKE_THREAD_LIBS_INIT})
