--- src/groveloudness/CMakeLists.txt
@@ -1,5 +1,0 @@
-set (libname "groveloudness")
-set (libdescription "upm grove loudness module")
-set (module_src ${libname}.cxx)
-set (module_h ${libname}.h)
-upm_module_init()
