--- src/groveloudness/javaupm_groveloudness.i
@@ -1,8 +1,0 @@
-%module javaupm_groveloudness
-%include "../upm.i"
-
-%{
-    #include "groveloudness.h"
-%}
-
-%include "groveloudness.h"
