--- src/groveloudness/jsupm_groveloudness.i
@@ -1,8 +1,0 @@
-%module jsupm_groveloudness
-%include "../upm.i"
-
-%{
-    #include "groveloudness.h"
-%}
-
-%include "groveloudness.h"
