--- src/groveloudness/pyupm_groveloudness.i
@@ -1,9 +1,0 @@
-%module pyupm_groveloudness
-%include "../upm.i"
-
-%feature("autodoc", "3");
-
-%include "groveloudness.h"
-%{
-    #include "groveloudness.h"
-%}
