--- src/hyld9767/jsupm_hyld9767.i
@@ -1,8 +1,0 @@
-%module jsupm_hyld9767
-%include "../upm.i"
-
-%{
-    #include "hyld9767.h"
-%}
-
-%include "hyld9767.h"
