+++ src/loudness/javaupm_loudness.i
@@ -1,0 +1,8 @@
+%module javaupm_loudness
+%include "../upm.i"
+
+%{
+    #include "loudness.h"
+%}
+
+%include "loudness.h"
