+++ src/loudness/jsupm_loudness.i
@@ -1,0 +1,8 @@
+%module jsupm_loudness
+%include "../upm.i"
+
+%{
+    #include "loudness.h"
+%}
+
+%include "loudness.h"
