+++ src/loudness/pyupm_loudness.i
@@ -1,0 +1,9 @@
+%module pyupm_loudness
+%include "../upm.i"
+
+%feature("autodoc", "3");
+
+%include "loudness.h"
+%{
+    #include "loudness.h"
+%}
